some settings updates, install script for *nix updated

+++ install.sh
@@ -1,15 +1,20 @@
 #!/bin/sh
 
-dir=~/code/dotfiles
+dir=${PWD}
 backupdir=dotfile_backup
 files="vimrc zshrc gitconfig zsh tmux.conf"
 
 cd $dir
-mkdir $backupdir
+
+if [ ! -d "$backupdir"]; then
+    mkdir $backupdir
+fi
 
 for file in $files; do
     echo "Copying existing files to $backupdir if they exist."
-    mv ~/.$file ./$backupdir/
+    if [ -f "$file" ]; then
+        mv ~/.$file ./$backupdir/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
 done
